requirements.txt
requests>=2.25.1
fastapi>=0.65.2
uvicorn>=0.14.0
pydantic>=1.8.2
httpx
websockets
numpy
geopy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]